bugfix the load rezistor and double parameter in computing the current

+++ firmware/powersupply/Inc/processing.h
@@ -3,5 +3,5 @@
 #include "micro.h"
 extern double Convert_to_Voltage(uint16_t ADC_LiniarVoltage);
 extern double ADC_to_Physical_VoltageDivider(double d_VoltageADC,uint32_t ui32RezistorUper,uint32_t ui32RezistorLower);
-extern double ADC_to_PhysicalINA169(uint16_t d_VoltageADC,double dRezitorSarcina,uint32_t ui32RezitorLeader);
+extern double ADC_to_PhysicalINA169(double d_VoltageADC,double dRezitorSarcina,uint32_t ui32RezitorLeader);
 #endif
